backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
flask-cors
flask-login
python-dotenv
psycopg2-binary
psycopg
gunicorn==20.1.0
flake8
pytest